% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{evaluate_dashboard}
\alias{evaluate_dashboard}
\title{Optimierungs-Ergebnisse auswerten}
\usage{
evaluate_dashboard(res, students, courses)
}
\arguments{
\item{res}{Das Ergebnis-Objekt der Funktion optimize_courses()}

\item{students}{Der data.frame der Schueler}

\item{courses}{Der data.frame der Kurse}
}
\description{
Zeigt eine Zusammenfassung der Schuelerzufriedenheit und des Kursstatus im Terminal an.
}
