% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_data_path}
\alias{example_data_path}
\title{Beispieldaten-Pfad abrufen}
\usage{
example_data_path()
}
\value{
Einen Dateipfad (Character) zur Excel-Vorlage.
}
\description{
Diese Funktion liefert den absoluten Pfad zur mitgelieferten Beispieldaten-Datei
(.xlsx) des Pakets. Diese Datei enthält bereits die Reiter "Schüler" und "Kurse".
}
