% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_demo_excel}
\alias{generate_demo_excel}
\title{Demo-Excel Datei generieren}
\usage{
generate_demo_excel(path)
}
\arguments{
\item{path}{Zielpfad}
}
\description{
Demo-Excel Datei generieren
}
