% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{get_demo_paths}
\alias{get_demo_paths}
\title{Pfade zu den Demo-Daten abrufen}
\usage{
get_demo_paths()
}
\value{
Der Pfad zur Gesamt-Excel-Vorlage (.xlsx)
}
\description{
Pfade zu den Demo-Daten abrufen
}
