% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{get_score}
\alias{get_score}
\title{Gibt den Matrixwert fuer die Score Matrix zurueck}
\usage{
get_score(score_matrix, i, j)
}
\description{
Gibt den Matrixwert fuer die Score Matrix zurueck
}
\keyword{internal}
