% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_data}
\alias{import_data}
\title{Daten importieren und validieren}
\usage{
import_data(students_file, courses_file)
}
\arguments{
\item{students_file}{Pfad zur CSV-Datei mit Schuelerdaten.}

\item{courses_file}{Pfad zur CSV-Datei mit Kursdaten.}
}
\value{
Eine Liste mit zwei Elementen: \code{students} und \code{courses} als data.frames.
}
\description{
Liest Schueler- und Kursdaten aus CSV-Dateien ein und validiert die Struktur rigoros.
Falls kein \code{tie_breaker} vorhanden ist, wird dieser automatisch generiert.
}
