% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{launch_gui}
\alias{launch_gui}
\title{Kurszuweisung GUI starten}
\usage{
launch_gui(
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\arguments{
\item{port}{Der Port, auf dem die App laufen soll (Standard: Zufällig).}

\item{launch.browser}{Logisch, ob die App im Browser geöffnet werden soll (Standard: TRUE).}
}
\description{
Startet eine interaktive Web-Oberfläche (Shiny) zur Optimierung der Kurswahlen.
Die GUI erlaubt den Datei-Upload, die Parametrisierung und die Visualisierung
der Ergebnisse.
}
