% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nil_coalescing}
\alias{nil_coalescing}
\alias{\%||\%}
\title{Safe NULL coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The value to check}

\item{b}{The value to return if a is NULL}
}
\value{
a if not NULL, else b
}
\description{
Safe NULL coalescing operator
}
