% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_courses}
\alias{optimize_courses}
\title{Kurswahl optimieren}
\usage{
optimize_courses(
  students,
  courses,
  enforce_survival = FALSE,
  balance_gender = 0,
  balance_class = 0,
  time_limit_sec = 15
)
}
\arguments{
\item{students}{data.frame mit Schueler-Praeferenzen}

\item{courses}{data.frame mit Kurs-Kapazitaeten}

\item{enforce_survival}{Logisch, ob Kurse bevorzugt gerettet werden sollen. Standard ist FALSE.}

\item{balance_gender}{Numerisch (0-100), Staerke der Geschlechter-Gleichverteilung. Standard ist 0.}

\item{balance_class}{Numerisch (0-100), Staerke der Klassen-Durchmischung. Standard ist 0.}

\item{time_limit_sec}{Zeitlimit fuer den Solver in Sekunden. Standard ist 15.}
}
\value{
Ein Listen-Objekt mit den Zuweisungen und Status
}
\description{
Berechnet die optimale Zuweisung basierend auf Schueler-Wünschen und Kurs-Kapazitäten.
Schützt automatisch vor Mehrfachwahlen desselben Kurses durch einen Schueler.
}
