% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pkg_colors}
\alias{pkg_colors}
\title{Standard-Farben fuer Visualisierungen}
\format{
An object of class \code{list} of length 6.
}
\usage{
pkg_colors
}
\description{
Standard-Farben fuer Visualisierungen
}
\keyword{datasets}
