% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_course_occupancy}
\alias{plot_course_occupancy}
\title{Kursauslastung visualisieren}
\usage{
plot_course_occupancy(eval_res)
}
\arguments{
\item{eval_res}{Resultat von evaluate_dashboard()}
}
\value{
Ein plotly Objekt
}
\description{
Kursauslastung visualisieren
}
