% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_satisfaction_demographics}
\alias{plot_satisfaction_demographics}
\title{Zufriedenheit nach Geschlecht visualisieren}
\usage{
plot_satisfaction_demographics(eval_res, total_m, total_w)
}
\arguments{
\item{eval_res}{Resultat von evaluate_dashboard()}

\item{total_m}{Anzahl Jungen gesamt}

\item{total_w}{Anzahl Maedchen gesamt}
}
\value{
Ein plotly Objekt
}
\description{
Zufriedenheit nach Geschlecht visualisieren
}
