% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{evaluate_dashboard}
\alias{evaluate_dashboard}
\title{Optimierungs-Ergebnisse auswerten}
\usage{
evaluate_dashboard(res, students, courses)
}
\arguments{
\item{res}{Das Ergebnis-Objekt der Funktion optimize_courses()}

\item{students}{Der data.frame der Schueler}

\item{courses}{Der data.frame der Kurse}
}
\value{
Eine unsichtbare Liste mit:
\item{choice_1, choice_2, choice_3}{Anzahl der Schueler mit jeweiliger Wahl}
\item{unassigned}{Anzahl der nicht zugewiesenen Schueler}
\item{course_stats}{data.frame mit Teilnehmerzahlen und Gesamtnachfrage pro Kurs}
\item{rest_students}{data.frame der nicht zugewiesenen Schueler}
}
\description{
Zeigt eine Zusammenfassung der Schuelerzufriedenheit und des Kursstatus im Terminal an.
Berechnet zudem detaillierte Statistiken fuer die Darstellung im Dashboard.
}
