% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_data}
\alias{import_data}
\title{Daten importieren und validieren}
\usage{
import_data(file_path)
}
\arguments{
\item{file_path}{Pfad zur Excel-Datei (.xlsx).}
}
\value{
Eine Liste mit zwei Elementen: \code{students} und \code{courses} als data.frames.
}
\description{
Liest Schueler- und Kursdaten aus einer Excel-Datei (.xlsx) ein.
Die Datei muss zwei Reiter haben: "Schüler" und "Kurse".
Falls kein \code{tie_breaker} vorhanden ist, wird dieser automatisch generiert.
}
